hotBound-backend/requirements.txt
flask==3.0.3
google-generativeai==0.8.3
deepgram-sdk==2.4.0
python-dotenv==0.19.0
Werkzeug==3.0.4
chromadb==0.5.15
